refactor: refactor directory hierarchy
